#192 Version checks on ProduceManagerRepository, etc.
Updated some AuthRepository methods signing the user out when version mismatch when it is not supposed too.

Updated write methods of ProduceManagerRepository to have app checks

+++ lib/features/produce_manager/data/repository/produce_manager_repository.dart
@@ -17,6 +17,7 @@ import 'package:farmhub/features/produce_manager/domain/entities/price/price.dar
 import 'package:farmhub/features/produce_manager/domain/entities/produce/produce.dart';
 import 'package:farmhub/features/produce_manager/domain/i_produce_manager_repository.dart';
 
+import '../../../../core/app_version_helper/app_version_helper.dart';
 import '../../../../core/auth/data/repository/auth_repository.dart';
 
 const String ProduceManagerRepositoryCode = "PMR-";
@@ -119,6 +120,16 @@ class ProduceManagerRepository implements IProduceManagerRepository {
     // Check if user is connected to the internet.
     if (await networkInfo.isConnected) {
       try {
+        // Check the user's app version
+        bool isAllowed = await AppVersionHelper.isAppVersionAllowed();
+        if (!isAllowed) {
+          return Left(AuthFailure(
+            code: ERR_APP_VERSION_NOT_SUPPORTED,
+            message: MESSAGE_APP_VERSION_NOT_SUPPORTED,
+            stackTrace: StackTrace.current,
+          ));
+        }
+
         // Check if user is signed in.
         final failureOrUserData =
             await authRepository.retrieveUserData().then((result) => result.fold(
@@ -214,6 +225,16 @@ class ProduceManagerRepository implements IProduceManagerRepository {
   }) async {
     if (await networkInfo.isConnected) {
       try {
+        // Check the user's app version
+        bool isAllowed = await AppVersionHelper.isAppVersionAllowed();
+        if (!isAllowed) {
+          return Left(AuthFailure(
+            code: ERR_APP_VERSION_NOT_SUPPORTED,
+            message: MESSAGE_APP_VERSION_NOT_SUPPORTED,
+            stackTrace: StackTrace.current,
+          ));
+        }
+
         final result = await pricesRemoteDatasource.addNewPrice(
           produceId: produceId,
           currentPrice: currentPrice,
@@ -258,6 +279,16 @@ class ProduceManagerRepository implements IProduceManagerRepository {
   FutureEither<Unit> editProduce(String produceId, String newProduceName) async {
     if (await networkInfo.isConnected) {
       try {
+        // Check the user's app version
+        bool isAllowed = await AppVersionHelper.isAppVersionAllowed();
+        if (!isAllowed) {
+          return Left(AuthFailure(
+            code: ERR_APP_VERSION_NOT_SUPPORTED,
+            message: MESSAGE_APP_VERSION_NOT_SUPPORTED,
+            stackTrace: StackTrace.current,
+          ));
+        }
+
         await remoteDatasource.editProduce(produceId, newProduceName);
 
         return const Right(unit);
@@ -277,6 +308,16 @@ class ProduceManagerRepository implements IProduceManagerRepository {
   FutureEither<Unit> deleteProduce(String produceId) async {
     if (await networkInfo.isConnected) {
       try {
+        // Check the user's app version
+        bool isAllowed = await AppVersionHelper.isAppVersionAllowed();
+        if (!isAllowed) {
+          return Left(AuthFailure(
+            code: ERR_APP_VERSION_NOT_SUPPORTED,
+            message: MESSAGE_APP_VERSION_NOT_SUPPORTED,
+            stackTrace: StackTrace.current,
+          ));
+        }
+
         await remoteDatasource.deleteProduce(produceId);
 
         return const Right(unit);
@@ -393,6 +434,16 @@ class ProduceManagerRepository implements IProduceManagerRepository {
   ) async {
     if (await networkInfo.isConnected) {
       try {
+        // Check the user's app version
+        bool isAllowed = await AppVersionHelper.isAppVersionAllowed();
+        if (!isAllowed) {
+          return Left(AuthFailure(
+            code: ERR_APP_VERSION_NOT_SUPPORTED,
+            message: MESSAGE_APP_VERSION_NOT_SUPPORTED,
+            stackTrace: StackTrace.current,
+          ));
+        }
+
         final result = await pricesRemoteDatasource.editSubPrice(
           produceId,
           priceId,
@@ -437,6 +488,16 @@ class ProduceManagerRepository implements IProduceManagerRepository {
       {required String produceId, required String priceId, required String subPriceDate}) async {
     if (await networkInfo.isConnected) {
       try {
+        // Check the user's app version
+        bool isAllowed = await AppVersionHelper.isAppVersionAllowed();
+        if (!isAllowed) {
+          return Left(AuthFailure(
+            code: ERR_APP_VERSION_NOT_SUPPORTED,
+            message: MESSAGE_APP_VERSION_NOT_SUPPORTED,
+            stackTrace: StackTrace.current,
+          ));
+        }
+
         final isPriceDocDeleted =
             await pricesRemoteDatasource.deleteSubPrice(produceId, priceId, subPriceDate);
         return Right(isPriceDocDeleted);
@@ -480,6 +541,16 @@ class ProduceManagerRepository implements IProduceManagerRepository {
   FutureEither<FarmhubUser> addToFavorites(FarmhubUser farmhubUser, String produceId) async {
     if (await networkInfo.isConnected) {
       try {
+        // Check the user's app version
+        bool isAllowed = await AppVersionHelper.isAppVersionAllowed();
+        if (!isAllowed) {
+          return Left(AuthFailure(
+            code: ERR_APP_VERSION_NOT_SUPPORTED,
+            message: MESSAGE_APP_VERSION_NOT_SUPPORTED,
+            stackTrace: StackTrace.current,
+          ));
+        }
+
         final updatedFarmhubUser = await remoteDatasource.addToFavorites(farmhubUser, produceId);
         globalAuthCubit.updateFarmhubUser(updatedFarmhubUser);
 
@@ -503,6 +574,16 @@ class ProduceManagerRepository implements IProduceManagerRepository {
   FutureEither<FarmhubUser> removeFromFavorites(FarmhubUser farmhubUser, String produceId) async {
     if (await networkInfo.isConnected) {
       try {
+        // Check the user's app version
+        bool isAllowed = await AppVersionHelper.isAppVersionAllowed();
+        if (!isAllowed) {
+          return Left(AuthFailure(
+            code: ERR_APP_VERSION_NOT_SUPPORTED,
+            message: MESSAGE_APP_VERSION_NOT_SUPPORTED,
+            stackTrace: StackTrace.current,
+          ));
+        }
+
         final updatedFarmhubUser =
             await remoteDatasource.removeFromFavorites(farmhubUser, produceId);
         globalAuthCubit.updateFarmhubUser(updatedFarmhubUser);
